Extracted mergeDefaults from src/utils/misc.tsx to src/utils/mergeDefaults.ts
