+++ src/utils/misc.tsx
@@ -21,25 +21,6 @@ import { Clipboard, Toasts } from "@webpack/common";
 import { DevsById } from "./constants";
 
 /**
- * Recursively merges defaults into an object and returns the same object
- * @param obj Object
- * @param defaults Defaults
- * @returns obj
- */
-export function mergeDefaults<T>(obj: T, defaults: T): T {
-    for (const key in defaults) {
-        const v = defaults[key];
-        if (typeof v === "object" && !Array.isArray(v)) {
-            obj[key] ??= {} as any;
-            mergeDefaults(obj[key], v);
-        } else {
-            obj[key] ??= v;
-        }
-    }
-    return obj;
-}
-
-/**
  * Calls .join(" ") on the arguments
  * classes("one", "two") => "one two"
  */
